in OTA for BG600, proceed to next steps only if it has successfully downloaded file localy

+++ insighioNode/apps/demo_console/ota.py
@@ -121,8 +121,9 @@ def downloadOTA(client, fileId, fileType, fileSize):
         if file_downloaded:
             local_file_name = device_info.get_device_root_folder() + file
             is_file_locally = client.modem_instance.get_file(file, local_file_name)
-            client.modem_instance.delete_file(file)
-            return local_file_name
+            if is_file_locally:
+                client.modem_instance.delete_file(file)
+                return local_file_name
         return None
     else:
         wCli = microWebCli.MicroWebCli(URL)
